purge after next build so any 3rd party components are taken into account
