+++ docs/pages/docs/configuration.mdx
@@ -86,6 +86,18 @@ Add `.html` to the end of each internal URL.
 }
 ```
 
+### PurgeCSS
+
+By default `next-ignite` will purge CSS it doesn't use.
+
+You can turn off this behavior:
+
+```json5
+{
+  "purge": false,
+}
+```
+
 ## Files
 
 ## `docs/public/logo.png`
